Move combobox components into `@/registry`
